refactor voltage initializer module
