+++ pypowsybl/voltage_initializer/__init__.py
@@ -4,10 +4,6 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 
-from .impl.voltage_initializer import (
-    run,
-    VoltageInitializerParameters,
-    VoltageInitializerResults,
-    VoltageInitializerObjective,
-    VoltageInitializerStatus
-)
+from .impl.voltage_initializer import run
+from .impl.voltage_initializer_results import VoltageInitializerResults, VoltageInitializerStatus
+from .impl.voltage_initializer_parameters import VoltageInitializerParameters, VoltageInitializerObjective
